+++ src/providers/podman/index.ts
@@ -1,5 +1,5 @@
 import { PodmanClient, initClient } from "./podmanClient";
-import { genBootnodeDef, genNodeDef } from "./dynResourceDefinition";
+import { genBootnodeDef, genNodeDef, replaceMultiAddresReferences } from "./dynResourceDefinition";
 import { setupChainSpec, getChainSpecRaw } from "./chain-spec";
 
 export const provider = {
@@ -9,4 +9,5 @@ export const provider = {
   initClient,
   setupChainSpec,
   getChainSpecRaw,
+  replaceMultiAddresReferences,
 };
